+++ src/RequestBuilder/Enums/LogicOperators.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PostgrestPhp\RequestBuilder\Enums;
+
+enum LogicOperators: string
+{
+    case NOT = 'not';
+    case OR = 'or';
+    case AND = 'and';
+}